% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ini_parse.R
\name{ini_parse}
\alias{ini_parse}
\title{Parse a .ini or a .cfg file}
\usage{
ini_parse(x)
}
\arguments{
\item{x}{File paths, one or more}
}
\description{
Parse a .ini or a .cfg file
}
\examples{
\dontrun{
file1 <- system.file("examples", "example1.ini", package="inir")
ini_parse(file1)

file2 <- system.file("examples", "example2.ini", package="inir")
ini_parse(c(file1, file2))
}
}

